Methods

Index

Main.KeyVNA.VNAParameters

Main.KeyVNA.clearStatus

Main.KeyVNA.complexFromTrace

Main.KeyVNA.connect

Main.KeyVNA.deleteAllTraces

Main.KeyVNA.deleteTrace

Main.KeyVNA.disconnect

Main.KeyVNA.getFrequencies

Main.KeyVNA.getSweepTime

Main.KeyVNA.getTrace

Main.KeyVNA.getTraceCatalog

Main.KeyVNA.getTraceFromMemory

Main.KeyVNA.identify

Main.KeyVNA.instrumentSimplifiedSetup

Main.KeyVNA.recv

Main.KeyVNA.saveS2P

Main.KeyVNA.send

Main.KeyVNA.setAveraging

Main.KeyVNA.setCalibration

Main.KeyVNA.setFastSweep

Main.KeyVNA.setFormat2Log

Main.KeyVNA.setFrequencies

Main.KeyVNA.setIFBandwidth

Main.KeyVNA.setMeasurement

Main.KeyVNA.setPowerLevel

Main.KeyVNA.setSweepPoints

Main.KeyVNA.setupFromFile

Main.KeyVNA.storeTraceInMemory

Main.KeyVNA.triggerContinuous

Main.KeyVNA.triggerHold

Main.KeyVNA.triggerSingle

Structs

Main.KeyVNA.VNAParameters — Type

VNAParameters

This struct is used to store some important VNA settings and is returned by instrumentSimplifiedSetup.

The fields are:

calName::String Calibration label

power::Integer Power level

center::Float64 Frequency center

span::Float64 Frequency span

ifbandwidth::Integer IF bandwidth

sweepPoints::Integer Sweep points

sweepTime::Float64 Sweep time

fastSweep::Bool Fast sweep

Functions

Main.KeyVNA.connect — Function

connect(host; port=5025)

Establish a TCP connection to the VNA with IP address host. Return a TCPSocket if succesful and nothing if not.

[!NOTE] nothing is a return type used in Julia for indicating, that it is intended to return nothing.

Main.KeyVNA.disconnect — Function

disconnect(socket)

Closes the TCP connection socket.

Main.KeyVNA.identify — Function

identify(socket)

Main.KeyVNA.clearStatus — Function

clearStatus(socket)

Main.KeyVNA.send — Function

send(socket::TCPSocket, msg::String)

Send msg to the TCP socket socket.

Can be used to send any SCPI command directly to the VNA.

Main.KeyVNA.recv — Function

recv(socket::TCPSocket)

Receive and return all buffered bytes from socket. If none are available, wait. Returns a Vector{UInt8}.

recv(socket::TCPSocket, nb::Integer)

Receive and return nb number of bytes from socket. Returns a Vector{UInt8}.

Main.KeyVNA.instrumentSimplifiedSetup — Function

instrumentSimplifiedSetup(socket::Sockets.TCPSocket; <keyword arguments>)

Change basic settings of the VNA specified by socket. Returns struct VNAParameters.

Keyword Arguments

calName::String = cals[:c3GHz]: Name of the calibration to use

power::Int = -20: Power level (dBm)

center::Float64 = 20.025e9: Center of the frequency band (Hz)

span::Float64 = 50e6: Span of the frequency band (Hz)

ifbandwidth::Int = Int(5e6): IF Bandwidth (Hz)

sweepPoints::Int = 101 Number of sweep points in frequency band

fastSweep::Bool = true: Enable Fast Sweep mode

measurement::String = "CH1_S11_1": Set which Channel, Parameter and Trace to measure

Main.KeyVNA.setPowerLevel — Function

setPowerLevel(socket::TCPSocket, power::Integer)

Set the power level, of the VNA specified by socket, to power dBm.

Main.KeyVNA.setCalibration — Function

setCalibration(socket::TCPSocket ,calName::String)

Set the calibration file, of the VNA specified by socket, to calName

Main.KeyVNA.setFrequencies — Function

setFrequencies(socket::TCPSocket, center::Float64, span::Float64)

Set frequency band, of the VNA specified by socket. The band is given by the center frequency and the span. The bounds are then given by center ± span/2.

Main.KeyVNA.setSweepPoints — Function

setSweepPoints(socket::TCPSocket, points::Integer)

Set number of sweep points, of the VNA specified by socket.

Main.KeyVNA.setIFBandwidth — Function

setIFBandwidth(socket::TCPSocket, bandwidth::Integer)

Set IF bandwidth, of the VNA specified by socket.

Main.KeyVNA.setMeasurement — Function

setMeasurement(socket::TCPSocket, name::String)

Set wich measurement, of the VNA specified by socket, to use.

name

The string has to be in a certain format "CH<n>_S<p>_<t>"`, with:

n: Channel number

p: Scattering parameter

t: trace number

For example the String "CH1_S11_1" selects Channel 1, the parameter S11 and the trace 1.

Main.KeyVNA.setFormat2Log — Function

setFormat2Log(socket::TCPSocket)

Set the format of the VNA display to logarithmic scale.

Main.KeyVNA.setFastSweep — Function

setFastSweep(socket::TCPSocket, fast::Bool)

Enable Fast Sweep mode.

Main.KeyVNA.setAveraging — Function

setAveraging(socket::TCPSocket, state::Bool; counts::Int=50)

TODO

Main.KeyVNA.setupFromFile — Function

setupFromFile(socket::TCPSocket,file::String)

Set up the VNA specified by socket according to the settings specified in file.

TODO: For more information see ...

Main.KeyVNA.triggerContinuous — Function

triggerContinuous(socket::TCPSocket)

Set the trigger of the VNA to continous.

Main.KeyVNA.triggerHold — Function

triggerHold(socket::TCPSocket)

Set the trigger of the VNA to hold.

Main.KeyVNA.triggerSingle — Function

triggerSingle(socket::TCPSocket)

Trigger the VNA once.

Main.KeyVNA.saveS2P — Function

saveS2P(socket::TCPSocket, fileURL::String)

Save the measurement as a S2P file fileUrl onto the VNA internal harddrive.

Note, that the file is not beeing saved on the VNA and not the device, where this function is being called. The file can then be transfered manually.

Main.KeyVNA.getFrequencies — Function

getFrequencies(socket::TCPSocket)

Return all the actual sweep points as a Vector{Float64}.

The sweep points are the frequencies at which the scattering parameter is measured during a sweep. They are specified by the number of sweeppoints and the frequency range.

See also setSweepPoints, setFrequencies.

Main.KeyVNA.getSweepTime — Function

getSweepTime(socket::TCPSocket)

Returns the time the VNA takes for the sweep.

Main.KeyVNA.getTrace — Function

getTrace(socket::TCPSocket)

Performs a sweep and returns a Vector{ComplexF64} of the scattering parameter.

Each element corresponds the the scattering parameter at the frequency given by the element of getFrequencies with the same index.

Main.KeyVNA.storeTraceInMemory — Function

storeTraceInMemory(socket::TCPSocket, mnum::Integer)

Perform a sweep and store the trace in the memory of the VNA.

mnum must be a unique integer. It is used to store and identify multiple traces at the same time. If a trace with the same mnum already exists, the VNA will throw an error and the trace is not stored. Existing traces can only be overridden if the it is deleted first. The trace is stored as a measurement called "data_<mnum>" on the VNA. The prefix "data_" is used to distinguish the stored traces from any other created by the VNA itself.

See also deleteTrace, deleteAllTraces and getTraceFromMemory.

Main.KeyVNA.getTraceFromMemory — Function

getTraceFromMemory(socket::TCPSocket, mnum::Integer; delete=true)

Returns the trace stored in the memory specified by mnum.

See also getTrace, deleteTrace and deleteAllTraces.

Main.KeyVNA.getTraceCatalog — Function

getTraceCatalog(socket::TCPSocket)

Returns a list as a String, with all traces stored in the VNA. Not only those, that are created by storeTraceInMemory.

Main.KeyVNA.deleteTrace — Function

deleteTrace(socket::TCPSocket, mnum::Integer)

Delete the trace with the identifier mnum.

Main.KeyVNA.deleteAllTraces — Function

deleteAllTraces(socket::TCPSocket)

Delete all traces, which were created by storeTraceInMemory.

Main.KeyVNA.complexFromTrace — Function

complexFromTrace(data::Vector{Float64})

Takes the raw datareturned by the VNA and returns aVector{ComplexF64}`.

The VNA returns a Vector{Float64}, where the each complex number is represented by two successive Float64, where the first is the real part and the second the imaginary.